Convert `popup=popup` to `popup=auto` or just `popup`
Per the [1] resolution, `popup=popup` has been renamed to `popup=auto`.
Additionally, per the discussion at [2], it is likely that "boolean-
like" behavior will also be adopted, such that `<div popup>` means
the same thing as `<div popup=auto>`. This CL implements both of
these changes.

Note that this CL has one case that still needs to be fixed:
  <div id=foo popup=invalid>
  <script>
    foo.popup === null; // false, should be true
  </script>

To fix the above, I need to figure out how to specify the ReflectMissing
and ReflectInvalid values so that they mean "null".

[1] https://github.com/openui/open-ui/issues/491#issuecomment-1132165724
[2] https://github.com/openui/open-ui/issues/533

Bug: 1307772
Change-Id: I6037c5322f7408ebd2c91690f89ecbc513c66bdb